requirements.txt
PySide6>=6.5
pyodbc>=4.0.35
python-dotenv>=1.0.0
